requirements.txt
pandas
qrcode
Pillow
reportlab